backfill_lambda/requirements.txt
supabase
playwright==1.40.0
python-dotenv
awslambdaric
